pyproject.toml
[tool.poetry]
name = "berserker"
version = "0.1.0"
description = "A discord bot written in python to handle voice channels in community discord servers."
authors = ["Akins <marleyakins@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.0.1"

[scripts]
berserker = "src:main"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
